pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amygdala"
version = "0.1.0"
description = "AI coding assistant memory system — git-integrated file summaries"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "sinanata" }]
keywords = ["ai", "memory", "coding-assistant", "git", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pydantic>=2.5",
    "typer>=0.12",
    "rich>=13",
    "httpx>=0.27",
    "mcp>=1.0",
    "jinja2>=3.1",
    "tomli-w>=1.0",
    "pyyaml>=6.0",
    "anyio>=4.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.39"]
openai = ["openai>=1.50"]
ollama = ["ollama>=0.4"]
all-providers = ["amygdala[anthropic,openai,ollama]"]
dev = [
    "pytest>=8",
    "pytest-cov>=6",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "coverage[toml]>=7.4",
    "mypy>=1.10",
    "ruff>=0.8",
    "respx>=0.21",
]

[project.scripts]
amygdala = "amygdala.cli.app:main"

[project.entry-points."amygdala.providers"]
anthropic = "amygdala.providers.anthropic:AnthropicProvider"
openai = "amygdala.providers.openai:OpenAIProvider"
ollama = "amygdala.providers.ollama:OllamaProvider"

[project.entry-points."amygdala.adapters"]
claude-code = "amygdala.adapters.claude_code.adapter:ClaudeCodeAdapter"

[tool.hatch.build.targets.wheel]
packages = ["src/amygdala"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--strict-markers -ra"
markers = [
    "integration: marks integration tests",
]

[tool.coverage.run]
source = ["amygdala"]
branch = true

[tool.coverage.report]
fail_under = 85
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "@overload",
    "raise NotImplementedError",
    "\\.\\.\\.",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
